Rename WinRT files to UWP
