+++ modules/freetype/uwpdef.h
@@ -1,5 +1,5 @@
 /*************************************************************************/
-/*  winrtdef.h                                                           */
+/*  uwpdef.h                                                             */
 /*************************************************************************/
 /*                       This file is part of:                           */
 /*                           GODOT ENGINE                                */
