+++ modules/webm/libvpx/SCsub
@@ -257,7 +257,7 @@ cpu_bits = env["bits"]
 osx_fat = (env["platform"] == 'osx' and cpu_bits == 'fat')
 webm_cpu_x86 = False
 webm_cpu_arm = False
-if env["platform"] == 'winrt':
+if env["platform"] == 'uwp':
     if 'arm' in env["PROGSUFFIX"]:
         webm_cpu_arm = True
     else:
@@ -306,7 +306,7 @@ if webm_cpu_x86:
         env_libvpx["ASFLAGS"] = '-I' + libvpx_dir[1:]
         env_libvpx["ASCOM"] = '$AS $ASFLAGS $TARGET $SOURCES'
     else:
-        if env["platform"] == 'windows' or env["platform"] == 'winrt':
+        if env["platform"] == 'windows' or env["platform"] == 'uwp':
             env_libvpx["ASFORMAT"] = 'win'
         elif env["platform"] == 'osx':
             env_libvpx["ASFORMAT"] = 'macho'
@@ -332,7 +332,7 @@ if webm_cpu_arm:
         env_libvpx["ASFLAGS"] = '-arch armv7'
     elif env["platform"] == 'android':
         env_libvpx["ASFLAGS"] = '-mfpu=neon'
-    elif env["platform"] == 'winrt':
+    elif env["platform"] == 'uwp':
         env_libvpx["AS"] = 'armasm'
         env_libvpx["ASFLAGS"] = ''
         env_libvpx["ASCOM"] = '$AS $ASFLAGS -o $TARGET $SOURCES'
@@ -382,7 +382,7 @@ elif webm_cpu_arm:
         env_libvpx_neon.Append(CCFLAGS=['-mfpu=neon'])
     env_libvpx_neon.add_source_files(env.modules_sources, libvpx_sources_arm_neon)
 
-    if env["platform"] == 'winrt':
+    if env["platform"] == 'uwp':
         env_libvpx.add_source_files(env.modules_sources, libvpx_sources_arm_neon_armasm_ms)
     elif env["platform"] == 'iphone':
         env_libvpx.add_source_files(env.modules_sources, libvpx_sources_arm_neon_gas_apple)
