+++ platform/uwp/SCsub
@@ -3,15 +3,15 @@
 Import('env')
 
 files = [
-    'thread_winrt.cpp',
+    'thread_uwp.cpp',
     '#platform/windows/tcp_server_winsock.cpp',
     '#platform/windows/packet_peer_udp_winsock.cpp',
     '#platform/windows/stream_peer_winsock.cpp',
     '#platform/windows/key_mapping_win.cpp',
-    'joystick_winrt.cpp',
+    'joystick_uwp.cpp',
     'gl_context_egl.cpp',
     'app.cpp',
-    'os_winrt.cpp',
+    'os_uwp.cpp',
 ]
 
 if "build_angle" in env and env["build_angle"]:
