+++ platform/uwp/app.cpp
