+++ platform/uwp/export/export.h
@@ -1,5 +1,5 @@
 /*************************************************************************/
-/*  thread_winrt.h                                                       */
+/*  export.h                                                             */
 /*************************************************************************/
 /*                       This file is part of:                           */
 /*                           GODOT ENGINE                                */
@@ -26,38 +26,4 @@
 /* TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE     */
 /* SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                */
 /*************************************************************************/
-#ifndef THREAD_WINRT_H
-#define THREAD_WINRT_H
-
-#ifdef WINRT_ENABLED
-
-#include "os/thread.h"
-
-#include <thread>
-
-class ThreadWinrt : public Thread {
-
-	std::thread thread;
-
-	static Thread* create_func_winrt(ThreadCreateCallback p_callback,void *,const Settings&);
-	static ID get_thread_ID_func_winrt();
-	static void wait_to_finish_func_winrt(Thread* p_thread);
-
-	ThreadWinrt();
-public:
-
-
-	virtual ID get_ID() const;
-
-	static void make_default();
-
-
-	~ThreadWinrt();
-
-};
-
-
-#endif
-
-#endif
-
+void register_uwp_exporter();
