+++ platform/uwp/joystick_uwp.h
@@ -1,5 +1,5 @@
 /*************************************************************************/
-/*  joystick.h                                                           */
+/*  joystick_uwp.h                                                       */
 /*************************************************************************/
 /*                       This file is part of:                           */
 /*                           GODOT ENGINE                                */
