+++ platform/uwp/logo.png
