+++ platform/uwp/thread_uwp.cpp
@@ -1,5 +1,5 @@
 /*************************************************************************/
-/*  export.h                                                             */
+/*  thread_uwp.cpp                                                       */
 /*************************************************************************/
 /*                       This file is part of:                           */
 /*                           GODOT ENGINE                                */
@@ -26,4 +26,49 @@
 /* TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE     */
 /* SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                */
 /*************************************************************************/
-void register_winrt_exporter();
+#include "thread_uwp.h"
+
+#include "os/memory.h"
+
+Thread* ThreadWinrt::create_func_winrt(ThreadCreateCallback p_callback,void *p_user,const Settings&) {
+
+	ThreadWinrt* thread = memnew(ThreadWinrt);
+
+
+	std::thread new_thread(p_callback, p_user);
+	std::swap(thread->thread, new_thread);
+
+	return thread;
+};
+
+Thread::ID ThreadWinrt::get_thread_ID_func_winrt() {
+
+	return std::hash<std::thread::id>()(std::this_thread::get_id());
+};
+
+void ThreadWinrt::wait_to_finish_func_winrt(Thread* p_thread) {
+
+	ThreadWinrt *tp=static_cast<ThreadWinrt*>(p_thread);
+	tp->thread.join();
+};
+
+
+Thread::ID ThreadWinrt::get_ID() const {
+
+	return std::hash<std::thread::id>()(thread.get_id());
+};
+
+void ThreadWinrt::make_default() {
+	create_func = create_func_winrt;
+	get_thread_ID_func = get_thread_ID_func_winrt;
+	wait_to_finish_func = wait_to_finish_func_winrt;
+};
+
+ThreadWinrt::ThreadWinrt() {
+
+};
+
+ThreadWinrt::~ThreadWinrt() {
+
+};
+
