+++ platform/uwp/thread_uwp.h
@@ -1,5 +1,5 @@
 /*************************************************************************/
-/*  thread_winrt.cpp                                                     */
+/*  thread_uwp.h                                                         */
 /*************************************************************************/
 /*                       This file is part of:                           */
 /*                           GODOT ENGINE                                */
@@ -26,49 +26,38 @@
 /* TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE     */
 /* SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                */
 /*************************************************************************/
-#include "thread_winrt.h"
+#ifndef THREAD_WINRT_H
+#define THREAD_WINRT_H
 
-#include "os/memory.h"
+#ifdef WINRT_ENABLED
 
-Thread* ThreadWinrt::create_func_winrt(ThreadCreateCallback p_callback,void *p_user,const Settings&) {
+#include "os/thread.h"
 
-	ThreadWinrt* thread = memnew(ThreadWinrt);
+#include <thread>
 
+class ThreadWinrt : public Thread {
 
-	std::thread new_thread(p_callback, p_user);
-	std::swap(thread->thread, new_thread);
+	std::thread thread;
 
-	return thread;
-};
+	static Thread* create_func_winrt(ThreadCreateCallback p_callback,void *,const Settings&);
+	static ID get_thread_ID_func_winrt();
+	static void wait_to_finish_func_winrt(Thread* p_thread);
 
-Thread::ID ThreadWinrt::get_thread_ID_func_winrt() {
-
-	return std::hash<std::thread::id>()(std::this_thread::get_id());
-};
-
-void ThreadWinrt::wait_to_finish_func_winrt(Thread* p_thread) {
-
-	ThreadWinrt *tp=static_cast<ThreadWinrt*>(p_thread);
-	tp->thread.join();
-};
+	ThreadWinrt();
+public:
 
 
-Thread::ID ThreadWinrt::get_ID() const {
+	virtual ID get_ID() const;
 
-	return std::hash<std::thread::id>()(thread.get_id());
-};
+	static void make_default();
 
-void ThreadWinrt::make_default() {
-	create_func = create_func_winrt;
-	get_thread_ID_func = get_thread_ID_func_winrt;
-	wait_to_finish_func = wait_to_finish_func_winrt;
-};
 
-ThreadWinrt::ThreadWinrt() {
+	~ThreadWinrt();
 
 };
 
-ThreadWinrt::~ThreadWinrt() {
 
-};
+#endif
+
+#endif
